arrange completed method to a separte file
