+++ package.json
@@ -36,6 +36,7 @@
     "webpack-dev-server": "^4.8.1"
   },
   "dependencies": {
+    "file-loader": "^6.2.0",
     "lodash": "^4.17.21"
   },
   "homepage": "https://github.com/bamsi/list-structure#readme"
