+++ src/index.js
@@ -5,7 +5,8 @@ import _ from "lodash";
 import './style.css';
 import { addListItem } from './modules/add.module.js';
 import { loadData } from './modules/load-data.module.js';
-import { update, completed } from './modules/update.module.js';
+import { update } from './modules/update.module.js';
+import { completed } from './modules/completed.module';
 import { removeItems } from './modules/remove.module.js';
 
 const newItem = document.querySelector('#add-list');
