+++ src/modules/update.module.js
@@ -9,12 +9,4 @@ function update(item) {
   writeLocalStorage(items);
 }
 
-function completed(index, status) {
-  const items = readLocalStorage();
-  const element = items.filter((i) => i.id === index)[0];
-  element.completed = status;
-  items[index - 1] = element;
-  writeLocalStorage(items);
-}
-
-export { update, completed };
+export {update};
