+++ webpack.config.js
@@ -21,6 +21,14 @@ module.exports = {
         test: /\.css$/i,
         use: ['style-loader', 'css-loader'],
       },
+      {
+        test: /\.(png|svg|jpg|jpeg|gif)$/i,
+        loader: 'file-loader',
+        options: {
+            name: 'img/[name].[ext]',
+            publicPath: '/'
+        }
+      },
     ],
   },
   devServer: {
